Add real-time ping display for regions
Sometimes geography alone is not enough for deciding which servers to
prefer, and you want to see how long it actually takes to reach a game
region on your network. This commit adds pinging functionality to help
with that.

Now each region in the list will have a ping display next to its code.

Internally, this has been achieved by changing the map of region states
to hold new region entry structures as values rather storing regions
directly (the region itself is a member of the entry structure now).
Each region entry holds a ping value that is passed to the region widget
render function.

To maintain these ping values, we start a thread that pings all the
regions' main servers and then sends the updates over the broadcast
channel. To avoid spamming the servers, ping checks are spaced out by
10-second intervals after each ping round completion.

We use a broadcast channel to avoid modification of the region states
map outside the main thread, which would be quite convoluted, which
would require us to handle the reference management and locking.
Instead, with a channel, we can process ping updates every time we paint
the UI (which happens on the main thread).

The reason for using a broadcast channel specifically over a regular
MPSC channel - the one we used for other things like modals - is to
allow us to have an additional receiver in the UI refresher thread. This
thread requests a repaint whenever ping updates are available, keeping
the UI fresh without requiring user input. Unfortunately, this seems to
produce a warning at the start due to frequent paint discards, but this
doesn't cause any serious issues and is negligible.

Although pinging may seem easy, it relies on the ICMP protocol,
requiring the use of OS-dependent APIs. Thanks to the great work of
`surge-ping` crate's developers, we don't have to deal with this
ourselves. We still however have to handle situations where such APIs
are unavailable for some reason. If that is the case, a warning modal
will be shown to the user at the start of the program.